3819. Rotate Non Negative Elements

+++ Problems/3834. Merge Adjacent Equal Elements/3834_merge_adjacent_equal_elements_cpp/main.cpp
@@ -36,7 +36,7 @@ void printArray(vector<long long> array) {
     cout << endl;
 }
 
-void test(vector<int> nums, vector<int> expected) {
+void test(vector<int> nums, vector<long long> expected) {
     cout << "Array: ";
     printArray(nums);
